+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Models exported to bucket are prefixed with their compute plan's ID
+
 ## [0.3.1] - 2021-08-25
 
 ### Added